refactor(codegen): extract bus runtime hooks into bus/runtime.rs (Round 3c)
4 fns lifted into BusRuntime<'ctx>: emit_bus_drain, emit_bus_queue_destroy,
emit_bus_register, emit_bus_register_shm_ring. These emit the runtime
glue around the bus queue (drain on yield, destroy at dissolve) plus
per-topic payload-type registration at main's bindings prelude.

Surface bumps to pub(crate):
- Cx fields: main_cooperative_pools, current_cooperative_pool
- ShmRingBindingInfo fields: payload_type_name, slot_count, shm_name,
  overflow

stdlib/time imports BusRuntime to keep the post-sleep bus-drain call
resolving.

Round 3 complete. codegen.rs: 35331 -> 35071 (-260 lines). Total Round
3: 35,331 -> 35,071 + bus/wire.rs 1663 + bus/dispatch.rs 854 +
bus/runtime.rs 209 = 2,726 lines lifted to bus/.
