+++ crates/hale-codegen/src/bus/mod.rs
@@ -1,4 +1,5 @@
 //! Bus codegen (publish, subscribe, wire, runtime). Round 3 of the codegen model-org refactor.
 
 pub(crate) mod dispatch;
+pub(crate) mod runtime;
 pub(crate) mod wire;
